backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.5.15",
    "aio-pika>=9.6.2",
    "aiohttp>=3.13.5",
    "alembic>=1.18.4",
    "assemblyai>=0.59.0",
    "beautifulsoup4>=4.14.3",
    "click>=8.3.1",
    "deepgram-sdk>=6.1.1",
    "fastapi>=0.135.3",
    "google-genai>=1.72.0",
    "lxml>=6.0.3",
    "openai>=2.30.0",
    "primp>=1.2.2",
    "psycopg2>=2.9.11",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "sqlalchemy>=2.0.49",
    "trafilatura>=2.0.0",
    "uvicorn>=0.42.0",
]
